+++ profiler/__main__.py
@@ -1,7 +1,7 @@
 # -*- coding: utf-8 -*-
 #
 # profiler : a Wi-Fi client capability analyzer tool
-# Copyright : (c) 2020-2021 Josh Schmelzle
+# Copyright : (c) 2024 Josh Schmelzle
 # License : BSD-3-Clause
 # Maintainer : josh@joshschmelzle.com
 
@@ -35,10 +35,10 @@ def init():
             "{0} only works on Linux... exiting...".format(os.path.basename(__file__))
         )
 
-    # hard set no support for python < v3.7
-    if sys.version_info < (3, 7):
+    # hard set no support for python < v3.9
+    if sys.version_info < (3, 9):
         sys.exit(
-            "{0} requires Python version 3.7 or higher...\nyou are trying to run with Python version {1}...\nexiting...".format(
+            "{0} requires Python version 3.9 or higher...\nyou are trying to run with Python version {1}...\nexiting...".format(
                 os.path.basename(__file__), platform.python_version()
             )
         )